+++ Dalaran/parser/__init__.py
@@ -1,1 +1,1 @@
-from .parser import *
+from .main import *
